pyproject.toml
[project]
name = "refrakt-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "boto3>=1.28.0",
    "pyyaml>=6.0",
    "openai>=1.0.0",
    "websockets>=12.0",
]
